added some util function, started file import handling
