+++ .env.local.example
@@ -1,0 +1,1 @@
+VITE_SERVER_ORIGIN="" # url your server is running on (ex. http://localhost:5000)
